requirements.txt
asgiref==3.7.2
Django==5.0.1
django-dbbackup==4.1.0
django-filter==23.5
django-rest-framework==0.1.0
django-simple-history==3.4.0
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
